boards: nrf9160_pca20035: Configure ADP536X in board_secure.c
Moves ADP536X setup to board_secure.c so that the board will
function properly for secure-only applications as well.
